implement xforwarded host and port headers
