+++ pilot/pkg/networking/core/httproute.go
@@ -23,13 +23,10 @@ import (
 
 	route "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
 	statefulsession "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/stateful_session/v3"
-	hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
 	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 	anypb "google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/durationpb"
-	wrappers "google.golang.org/protobuf/types/known/wrapperspb"
 
-	meshconfig "istio.io/api/mesh/v1alpha1"
 	networking "istio.io/api/networking/v1alpha3"
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
@@ -188,8 +185,10 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundHTTPRouteConfig(
 	util.SortVirtualHosts(virtualHosts)
 
 	if !useSniffing {
-		includeRequestAttemptCount := GetProxyHeaders(node, req.Push, istionetworking.ListenerClassSidecarOutbound).IncludeRequestAttemptCount
-		virtualHosts = append(virtualHosts, buildCatchAllVirtualHost(node, includeRequestAttemptCount))
+		ph := util.GetProxyHeaders(node, req.Push, istionetworking.ListenerClassSidecarOutbound)
+		appendXForwardedHost := ph.XForwardedHost
+		includeRequestAttemptCount := ph.IncludeRequestAttemptCount
+		virtualHosts = append(virtualHosts, buildCatchAllVirtualHost(node, includeRequestAttemptCount, appendXForwardedHost))
 	}
 
 	out := &route.RouteConfiguration{
@@ -291,72 +290,6 @@ func selectVirtualServices(virtualServices []config.Config, servicesByName map[h
 	return out
 }
 
-type ProxyHeaders struct {
-	ServerName                 string
-	ServerHeaderTransformation hcm.HttpConnectionManager_ServerHeaderTransformation
-	ForwardedClientCert        hcm.HttpConnectionManager_ForwardClientCertDetails
-	SetCurrentCertDetails      *meshconfig.ProxyConfig_ProxyHeaders_SetCurrentClientCertDetails
-	IncludeRequestAttemptCount bool
-	GenerateRequestID          *wrappers.BoolValue
-	SuppressDebugHeaders       bool
-	SkipIstioMXHeaders         bool
-	XForwardedPort             bool
-	XForwardedHost             bool
-}
-
-func GetProxyHeaders(node *model.Proxy, push *model.PushContext, class istionetworking.ListenerClass) ProxyHeaders {
-	pc := node.Metadata.ProxyConfigOrDefault(push.Mesh.DefaultConfig)
-	return GetProxyHeadersFromProxyConfig(pc, class)
-}
-
-func GetProxyHeadersFromProxyConfig(pc *meshconfig.ProxyConfig, class istionetworking.ListenerClass) ProxyHeaders {
-	base := ProxyHeaders{
-		ServerName:                 EnvoyServerName,
-		ServerHeaderTransformation: hcm.HttpConnectionManager_OVERWRITE,
-		ForwardedClientCert:        hcm.HttpConnectionManager_APPEND_FORWARD,
-		IncludeRequestAttemptCount: true,
-		SuppressDebugHeaders:       false,
-		GenerateRequestID:          nil, // Envoy default is to enable them, so set nil
-		SkipIstioMXHeaders:         false,
-		XForwardedPort:             false,
-		XForwardedHost:             false,
-	}
-	if class == istionetworking.ListenerClassSidecarOutbound {
-		// Likely due to a mistake, outbound uses "envoy" while inbound uses "istio-envoy". Bummer.
-		// We keep it for backwards compatibility.
-		base.ServerName = "" // Envoy default is "envoy" so no need to set it explicitly.
-	}
-	ph := pc.GetProxyHeaders()
-	if ph == nil {
-		return base
-	}
-	if ph.AttemptCount.GetDisabled().GetValue() {
-		base.IncludeRequestAttemptCount = false
-	}
-	if ph.ForwardedClientCert != meshconfig.ForwardClientCertDetails_UNDEFINED {
-		base.ForwardedClientCert = util.MeshConfigToEnvoyForwardClientCertDetails(ph.ForwardedClientCert)
-	}
-	if ph.Server != nil {
-		if ph.Server.Disabled.GetValue() {
-			base.ServerName = ""
-			base.ServerHeaderTransformation = hcm.HttpConnectionManager_PASS_THROUGH
-		} else if ph.Server.Value != "" {
-			base.ServerName = ph.Server.Value
-		}
-	}
-	if ph.RequestId.GetDisabled().GetValue() {
-		base.GenerateRequestID = proto.BoolFalse
-	}
-	if ph.EnvoyDebugHeaders.GetDisabled().GetValue() {
-		base.SuppressDebugHeaders = true
-	}
-	if ph.MetadataExchangeHeaders != nil && ph.MetadataExchangeHeaders.GetMode() == meshconfig.ProxyConfig_ProxyHeaders_IN_MESH {
-		base.SkipIstioMXHeaders = true
-	}
-	base.SetCurrentCertDetails = ph.SetCurrentClientCertDetails
-	return base
-}
-
 func BuildSidecarOutboundVirtualHosts(node *model.Proxy, push *model.PushContext,
 	routeName string,
 	listenerPort int,
@@ -387,7 +320,7 @@ func BuildSidecarOutboundVirtualHosts(node *model.Proxy, push *model.PushContext
 		listenerPort = 0
 	}
 
-	includeRequestAttemptCount := GetProxyHeaders(node, push, istionetworking.ListenerClassSidecarOutbound).IncludeRequestAttemptCount
+	includeRequestAttemptCount := util.GetProxyHeaders(node, push, istionetworking.ListenerClassSidecarOutbound).IncludeRequestAttemptCount
 
 	servicesByName := make(map[host.Name]*model.Service)
 	for _, svc := range services {
@@ -837,7 +770,7 @@ func getUniqueAndSharedDNSDomain(fqdnHostname, proxyDomain string) (partsUnique 
 	return
 }
 
-func buildCatchAllVirtualHost(node *model.Proxy, includeRequestAttemptCount bool) *route.VirtualHost {
+func buildCatchAllVirtualHost(node *model.Proxy, includeRequestAttemptCount bool, appendXForwardedHost bool) *route.VirtualHost {
 	if util.IsAllowAnyOutbound(node) {
 		egressCluster := util.PassthroughCluster
 		notimeout := durationpb.New(0)
@@ -856,7 +789,8 @@ func buildCatchAllVirtualHost(node *model.Proxy, includeRequestAttemptCount bool
 			Timeout: notimeout,
 			// Use deprecated value for now as the replacement MaxStreamDuration has some regressions.
 			// nolint: staticcheck
-			MaxGrpcTimeout: notimeout,
+			MaxGrpcTimeout:       notimeout,
+			AppendXForwardedHost: appendXForwardedHost,
 		}
 
 		return &route.VirtualHost{
