+++ pilot/pkg/networking/core/listener_inbound.go
@@ -808,7 +808,7 @@ func buildInboundBlackhole(lb *ListenerBuilder) *listener.FilterChain {
 
 // buildSidecarInboundHTTPOpts sets up HTTP options for a given chain.
 func buildSidecarInboundHTTPOpts(lb *ListenerBuilder, cc inboundChainConfig) *httpListenerOpts {
-	ph := GetProxyHeaders(lb.node, lb.push, istionetworking.ListenerClassSidecarInbound)
+	ph := util.GetProxyHeaders(lb.node, lb.push, istionetworking.ListenerClassSidecarInbound)
 	httpOpts := &httpListenerOpts{
 		routeConfig:      buildSidecarInboundHTTPRouteConfig(lb, cc),
 		rds:              "", // no RDS for inbound traffic
@@ -834,14 +834,6 @@ func buildSidecarInboundHTTPOpts(lb *ListenerBuilder, cc inboundChainConfig) *ht
 		port:                      int(cc.port.TargetPort),
 		statPrefix:                cc.StatPrefix(),
 		hbone:                     cc.hbone,
-	}
-
-	// Configure X-Forwarded-Port and X-Forwarded-Host headers
-	if ph.XForwardedPort {
-		httpOpts.connectionManager.XForwardedPort = proto.BoolTrue
-	}
-	if ph.XForwardedHost {
-		httpOpts.connectionManager.XForwardedHost = proto.BoolTrue
 	}
 
 	// See https://github.com/grpc/grpc-web/tree/master/net/grpc/gateway/examples/helloworld#configure-the-proxy
